Crate hyperdex

Macros

NewHyperMapAttribute!
Specifies a key-value pair for a particular map attribute.

NewHyperObject!
Creates a HyperDex object.

Structs

Admin
A HyperDex Admin object, used for meta operations like creating and removing spaces.

Client
A HyperDex client, used for common operations like getting and putting objects.

F64
A wrapper around f64.

HyperError
An error related to HyperDex.

HyperMapAttribute
A key-value pair associated with a specific map attribute

HyperObject
A HyperDex object.

HyperPredicate
A predicate used for search.

Enums

HyperObjectKeyError
The errors that can occur upon a lookup from a HyperObject.

HyperPredicateType
Predicates that HyperDex supports.

HyperValue
Types of values that HyperDex accepts.