Enum hyperdex::HyperObjectKeyError

pub enum HyperObjectKeyError {
    KeyDoesNotExist,
    ObjectIsAnotherType,
}

The errors that can occur upon a lookup from a HyperObject.

Variants

KeyDoesNotExist

The key does not exist.

ObjectIsAnotherType

The key does exist, but the value is not the type that you think it is.